fix: use bigint cast for by_tool token accumulation to prevent int overflow
The SQL COALESCE cast for daily_user_stats.by_tool tokens field used
::int (max ~2.1B) which silently truncated values for heavy users.
Changed to ::bigint. Also rebuilt by_tool data in production from
token_usage source of truth.

+++ apps/web/src/app/api/v1/sessions/route.ts
@@ -381,7 +381,7 @@ export async function POST(request: NextRequest) {
           byTool: sql`${dailyUserStats.byTool} || jsonb_build_object(
             ${sessionData.toolType}::text,
             jsonb_build_object(
-              'tokens', COALESCE((${dailyUserStats.byTool} -> ${sessionData.toolType}::text ->> 'tokens')::int, 0) + ${
+              'tokens', COALESCE((${dailyUserStats.byTool} -> ${sessionData.toolType}::text ->> 'tokens')::bigint, 0) + ${
                 sessionData.inputTokens +
                 sessionData.outputTokens +
                 (sessionData.cacheCreationTokens ?? 0) +
